Render the scene only for game states >= LOADING
This is what the regular GPU plugin does.

Instead of assuming that rendering is fine as long as the user is
logged in, we assume it is fine as long as the game state is some
value at or above LOADING, which seems to be the cutoff point between
logged-in states and logged-out states. The previous logic was to render
as long as the game state had been in the LOGGED_IN state before,
and hadn't yet transitioned back to the LOGIN_SCREEN state.