Switch to loading all lights from JSON, with support for hot-swapping
This commit converts the `lights.txt` config file, and the `NpcLight`,
`ObjectLight` and `ProjectileLight` enums into a single combined
`lights.json` file. In addition, the JSON file is continuously watched
through the file system, and changes are reflected live in-game.
Switching to JSON also paves the way to more easily implement dev tools
for testing different lighting configurations and making modifications
on-the-fly.

List of changes:
- New `ExportLightsToJson` class to convert from the old formats to JSON.
- Old enums and `lights.txt` are kept in the same locations temporarily
  to help with merging potential PRs that rely on old formats.
- The old `lights.txt` format parser has been moved out of the main
  source set. `lights.txt` itself and old enums remain for convenience
  with porting older PRs.
- The new `lights.json` file is now watched for changes through the file
  system and automatically reloaded when modified.
- Hot-swapping can be enabled when running the plugin normally by
  setting the environment variable `RLHD_LIGHTS_PATH` to the path of an
  alternate JSON file to load. The path can be either absolute or
  relative to the RuneLite process' current working directory.